+++ .eslintrc.js
@@ -1,7 +1,7 @@
 module.exports = {
   root: true,
-  // This tells ESLint to load the config from the package `@custom-lb-ui/eslint-config`
-  extends: ["@custom-lb-ui/eslint-config"],
+  // This tells ESLint to load the config from the package `@custom-lb/eslint-config`
+  extends: ["@custom-lb/eslint-config"],
   settings: {
     next: {
       rootDir: ["apps/*/"],